moving eligibility.test.js file and adding tests for more test coverage
